requirements.txt
# FleetFuel360 Requirements
# Python 3.10+ dependencies

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7

# Data Processing
pandas==2.1.0
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Data Visualization (for notebooks)
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter Notebooks
jupyter==1.0.0
ipykernel==6.25.1

# Development Tools
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
